refactor(scraper): move parsers to scraper module
